functions_nutrition/requirements.txt
azure-functions
azure-storage-blob
pandas
azure-cosmos